make Derive.Call.Idiom.String more powerful
In addition to a general refactor, it will now release a string based on
the next time the same string is used, not the next note, so it can now
handle polyphonic strings, in addition to a more consistent result for
monophonic ones.  Also it adds a string env var, so it can do general
purpose string selection selection.