change event ignore prefix from -- to --|
Since I'm getting rid of the `n` call, I can no longer use `n --` as an
explicit note.  I should be able to use plain `--`, but since the "ignore
event" feature is still useful, I moved it to `--|`.  This is more
orthogonal anyway, and in the past it was annoying that I couldn't have a
note with a comment.