switch from Scopes using list-based [LookupCall c] to map-based CallMap c
With this change I had to make a separate Builtins type with the "compiled"
version of Library.  The difference is that Library can have duplicates, and
compiling to Builtins means indexing by Module and warning about duplicate
names.

This is another big annoying change, but I think the new structure makes more
sense.  It should also be more efficient for imports, and especially if I want
to look up individual symbols within their modules, which I plan on doing when
I add qualified symbol support.  The main reason I had [LookupCall c] before
was because I originally didn't have modules at all.

One side-effect is that previously named calls could shadow pattern calls by
being ahead of them in the list, but now pattern calls are always behind named
ones.  Which is maybe the other reason for [LookupCall c]: that way an import
can always shadow existing stuff.  I don't think this is a problem because I
think all the PatternCalls are in the prelude anyway, and maybe if I do import
them I still want them to be lower priority, but if it turns out to be a
problem I might have to go to [CallMap c] or something.