add qualified symbol syntax, and update the docs
That was a huge amount of preparatory work for a simple feature, and even one
I'm not totally convinced is super valuable.

Still, I think the work made scopes and modules less weird and probably more
efficient, so maybe it was just technical debt that would have needed to have
been paid at some point anyway.