solkattu: change Sequence._marks from Maybe (GroupMark g) to [GroupMark g]
This way Sequence.flatten supports nested groups and doesn't discard all
but one.  The higher levels like Realize still just take the first one
though, so the result is the same.

The main problem is that I'm still not sure exactly how to treat nested
groups, but at least I have a lot of notes in the TODO now, and a plan for
how to at least make 'sandi' not break in the presence of 'reduce'.