solkattu: rework groups, mostly to make nested groups work
This wound up being way to big and complicated, and it's not even done yet.
Every time I tried one approach, various things would break or get too
complicated.

The essential change is that previously groups were stored as the dropped
sollus, which were then put back into the sequence by realize.  But this
destroyed any nested group information, which then turned out to be too
hard or impossible to reconstruct by the time it got to realize.  So the
new approach is that no sollus are dropped at all, and only the amount to
drop is stored.  Now realize no longer needs to do any reconstruction, but
instead it needs to strip out the "extra" strokes in a second pass.  In
addition, the various matrasOf and durationOf functions now need to
subtract out the "dropped" sollus, since they're not inline with the rest.

Along the way, I changed the (meta, sollu) to a Flat type, since the former
couldn't represent nested groups which started in the same place, and
further changed group to a separate constructor since otherwise I can't
tell the difference between group (tempo ...) and tempo (group ...).

I also added an FMatra type for fractional Matra, and removed the various
splitD Duration-level functions in favor of native matra-level ones.
Duration, being absolute, is only valid at the top level of the sollu
expression, which makes it confusing and error-prone.

Even after all this, at least one Korvai is buggy and I think nested groups
*still* aren't handled correctly, but this patch has already dragged on way
to long already.