restore 0 and 1 limits for 'd' and 'u' control calls
I removed them in the "make the limit on the 'u' and 'd' control calls
optional" patch because the 1 limit made 'u' not useful in a tempo track.
But it turns out removing the limits leads to lots of confusing clipped
signal warnings from the MIDI performer.

I added EnvKey.control_gt_0 to solve the specific problem with tempo and
'u'.