linear: remove functions from Perform.Signal, including sig_min and sig_max
This means I removed the min and max track signal mergers as well, but I
don't think I ever used them.

The fewer things in Perform.Signal, the fewer I need to port to the linear
segment version.  sig_min and sig_max aren't impossible to implement for
linear, but could be annoying and no reason to do it until I need it.