linear: remove Score.event_control_offset and event_untransformed_*
When I switch to Segment.Signal, I'll have the offset built-in.

Recording this separately means a less cluttered switch to segments patch,
and also gives me an opportunity to profile and see if it ever was actually
worth it at all.  If it turns out to be completely not worth it maybe
Segment.Signal doesn't need the offset either.

The implementation cost for Score.event_control_offset wasn't that bad,
it just came down to about 60 lines in Derive.Score and then having to
type event_transformed_* or event_untransformed_*, but having to do that
always bothered me.  An optimization shouldn't clutter the API so much.