make TimeVector.constant put its sample at -large rather than 0
This change is already present in Util.Segment, so this is just a
pre-emptive change to fix a bug.  TimeVector has a hack where a
single sample <=0 means a constant signal.  This turns out to be brittle
because moving the event in a positive direction makes it no longer
constant.  This used to be hidden by Score.event_control_offset, and in
the future will once again be hidden by Segment._offset.

But clearly the whole thing about using a sample at 0 is a brittle hack.
Maybe I should use -Infinity?  Otherwise I might need a whole new
constructor.