linear: update Block.trim_controls to use drop_discontinuity_at
This winds up being exactly the same, but I put the functionality in its
own function and wrote one for the segment implementation.

trim_controls itself is a grody hack, and I spent quite a while trying to
think of a way to solve the problem better.  Having signals with
orientation might be such a way, but it's an orthogonal change, so as long
as I can preserve the hack, I'll put off doing it I can get the linear
segment stuff done with.