switch the signal implementation from piecewise-constant to piecewise-linear
Previously, a signal was a vector of (x, y) samples, where each y value is
constant until the next sample.  Now, each sample is the end of a linear
segment.

The reason for the change is that while individual constant samples are
enough to describe a smooth curve for low-bandwidth MIDI, I would need much
denser samples for non-MIDI instruments.  Ultimately this is because MIDI
synthesizers implicitly smooth out the control values, at the price of
latency.  Of course I could implement the same compromise for non-MIDI
instruments, but part of the point of developing non-MIDI instruments in
the first place is to avoid its limitations.

In fact, the Im backend was already assuming linear interpolation between
samples, but the inconsistent treatment was leading to bugs and awkward
API.

This was a surprisingly large change.  The greatest difference is that
constant samples can be sliced at arbitrary spots and most all
transformations can simply be applied directly to the y values.  But linear
segments need both ends defined to tell their slope, and only linear
transformations are valid, otherwise the segment has to be resampled.
Fortunately, most of my transformations are simple linear ones like (+) and
(*), and I use a lot of linear interpolation, so this is hopefully a fairly
natural representation.  Another unexpected result is that previously I
could slice a signal by dropping samples >=x, but since linear segments are
defined continuously, I have to keep the sample at x, and whoever

Of course MIDI still needs constant samples as always, so now the linear
segments are interpolated in the MIDI performer.

This patch is also much too large.  Even though I prepared as much as
possible beforehand, I couldn't make the APIs totally compatible, so it
still required a big switch.  I ported most all of the signal handling
calls, except a few which I disabled to port later: Rearticulate.slur, and
SignalTransform.c_quantize.

I also had to revise how SignalTransform.segment works so it can treat
linear interpolation specially, and I moved it along with the smoothing
functions to ControlUtil and PitchUtil, as appropriate.

In addition, I got rid of the feature where a signal with a single x at 0
is implicitly constant for all time.  The special cases were awkward and
lead to bugs.  Now constant signals just start at a large negative number.
I actually did most of the changes in a previous patch, but finally removed
the last vestiges here.

Another related change is that Warp is now implemented with composed
functions backed by a signal, rather than composing signals.  This let me
avoid having to implement Signal.compose and Signal.inverse_at for the
more complicated linear segments.

Also I incorporated an x-offset into the Signal type so I can translate
them for free.  This is a bit orthogonal so I probably should have left it
for later.  Or maybe I don't need it at all.

This should resolve all of the NOTE [signal-discontinuity]s that are
scattered around, or at least once I get things cleaned up.  Speaking of
which, I accumulated a large number of additional TODOs making the change,
which I'll be fixing later, but at least all the tests are passing again.
I expect all of the verify_performances to fail, because MIDI output has
surely changed for all scores.  I'll have to check and listen to make sure
it's still ok.