gamakam: extend the gamakam sequence with gamakam-above, which expects pitch track above the note track
The prev pitch was broken by the linear segments change, and it never
really worked reliably anyway, and was too complicated.  By putting pitch
above it gets a lot simpler, and if it turns out to work in all cases and
not have any fatal problems I haven't thought of, I'll switch to that way.