audio: fix a bug where Audio.synchronizeToSize didn't actually do that
There were two bugs: one is that it used chunkSize instead of the passed
size, and the other was that it turns out S.breakWhen breaks before the
condition becomes true, not after.

This was especially bad because later C assumed chunks had equal size, and
C doesn't check array bounds.