switch to using cpphs instead of whatever cpp is installed
This means I can finally use backslash string gaps in CPP-using modules.
The combination of the #ifdef TESTING convention and Derive.C.* including
inline documentation means I run into this problem.  Not super frequently,
but enough to be annoying.

Actually there was just one Derive.C.* module with that problem.  But I
know I've not bothered to put on export lists because that would imply
just get rid of the problem.