faust: add the ability to save and restore state
I'm not using it yet though.  This wound up being way more work than it
should have been, because faust doesn't export the the state from the C++
class, and I can't just serialize the object directly since it has a
vtable.  So I switched to faust C output, and rewrap it up in a Patch type.

As a bonus I can avoid allocating state for a patch if I just want query
it.