test: incorporate test/check_output into RunTests
This is in support of making a cabal port of the test framework, but I
figured I should make the changes here first, to avoid them drifting too
far from each other.

Ideally of course they could be the same, but I don't want to have to deal
with cabal any more than absolutely necessary.