make Pretty float instances extend precision until a non-zero
Otherwise a lot of numbers close to 0 just round to 0, which makes
pretty-printed data misleading when numbers are close to integers.

Something like 3.50005 will still show as 3.5, but there's no general
solution to that short of showing with full precision, and in practice
many test values are close to 0, not so many close to some arbitrary other
number.