im-incremental: add Sample.predictDuration
This uses the mystical quadratic formula to predict how long as sample
should be after resampling.  Unfortunately, while in simple cases it seems
to pretty closely match what libsamplerate actually does, in some extreme
cases they are very different.  I haven't yet figured out if this is a
problem with libsamplerate, or with the way I'm using it.