solkattu: invert pattern notation from p6 to 6p
This is consistent with all the other abstractions, which always put the
duration first.  Also, this is less ambiguous, since notation never uses
numbers, but 'p' is a common stroke name.

However, it's inconsistent with the Haskell and tracklang notation, due
to Haskell syntax.  Win some lose some I guess.