solkattu: allow rests in sollu maps, but require they match with strokes
Previously, I stripped rests from the sollus, and then paired them with
strokes, which might include rests.  This means that I'd have to write out
the strokes without rests and the rhythms would be taken only from the
sollus.  As strokes are now a full Sequence instead of just a list to
support sarvalaghu, they're getting more complicated and awkward to write
without rests.  Even in the shorter case, though, I think it's still
clearer to put rests in both sollus and strokes.