add Derive.Unset Merger
This is like Set, but biased to the old signal.  So it will add a signal
only if one doesn't already exist.

It turns out I don't really need it, since from within Haskell it's more
straightforward to do it explicitly with Derive.is_control_set.  But I
added it so I might as well keep it for the moment.