fix a bug where flags got deserialized incorrectly
This was caused by 1b37fd434e1c9b2cfce8943e5feb28c07813cec7.  The problem
was that using Serialize.get_enum and put_enum is not safe when the enum
is modified.  Ultimately the problem is that I stopped observing the
convention of putting Ui.State Serialize instances in Cmd.Serialize and
being careful about compatibility.  It would be nice to have a more
automatic way to enforce that.