move various instrument allocation functions to Cmd.Repl.LAlloc
Also, remove the rule that Dummy instrument allocations had to have
an empty Qualified.  It turns out they do inherit their Common.Common,
and in fact need to if I'm to have a place to put custom cmds and derive
calls and thru and whatnot.  It's a bit confusing that those things will
be MIDI (or whatever) instruments, but if it's too confusing I can add
a DummyInst someday, which is parallel to MidiInst and ImInst.