make instrument allocation settings an overlay on defaults instead of replacement
The idea is that it's a hassle when a patch change makes a bunch of scores
out of date, when they just want the default, but it turns out the main
culprit, patch tuning, is often only in the allocation after all.

But this also simplifies a bunch of places since they no longer need the
instrument db to copy over settings.