apply damage only to performances that have a the damage in their called set
This is a general optimization, but it was to make up for the loss of
the "notes not changed" optimization, which I had to get rid of to get
waveforms to initialize for the first time.

I also skip insert_damage if there is none, and accumulate damage strictly.
I'm not sure if that helps, but at the least it shouldn't hurt!