refactor Msg.ImStatus to put block and tracks in the parent type
Also I now filter out all but the root BlockId in Cmd.Performance, before
sending the Msgs, instead of in PlayC, which means I now do it for all
ImStatuses.

This originally menat to change how waveforms are loaded but it turns out
I already did that and it didn't work the first time either, but at least I
got this refactor out of it.