sampler: fix infinite loop introduced by c5d3e3eff6d7f330e6bf7e587872f6c5e303d792
Extending hashes to infinity in skipCheckpoints made a test diverge, but
that in turn turned up that not extending hashes in skipCheckpoints was
intentional, to avoid rerendering a final decay.  Figuring out how to
properly skip a decay turns out to be annoyingly subtle, but I think it
works now.