add a few optimizations
They are unsafeDupablePerformIO in Crc32Instances, memoized instrument
lookup, and Map.Strict in Midi.Perform.

It turns out they have no noticeable affect on the profiles, but I don't
entirely trust the profiles, and anyway it doesn't seem to hurt them.  I
tried using HashMap in Midi.Perform, and it hurt profiles a little bit in
addition to adding complexity, so I scrapped that.