faust: tweak controls so they move one control block earlier
Because controls have latency, I need to do this to make sure they reach
their intended values by the time the note starts, not right when it starts.
All the awkwardness makes me even more sure that faust's way of doing
controls is not the right way, compared to using audio inputs, but what can
be done?