audio: add Audio.Block, with a compact representation for constant signals
The idea is to use this to implement silences efficiently.  I haven't yet
pushed it all the way to a special file format, and I don't insert explicit
leading silence for faust instruments, so it probably doesn't improve
performance yet, but it should make it easier to implement that later.