fltk: re-implement EventTrack scrolling to cache the entire image
Long ago, I would draw the tracks incrementally.  However, I got lots of
drawing artifacts, probably due to uneven floating point roundoff, and
more generally due to it being pretty complicated to try to draw an
arbitrary part of the track perfectly.  So I gave up and redrew the whole
visible part of the track when scrolling.  This solved some problems, but I
still had artifacts when moving the selection since that was still
incremental.  And it made scrolling slow, which I tracked down to the low
level fltk draw_text() function, which in turn more or less directly
dispatches to the OS X text drawing.  Perhaps there are ways to to make that
faster, but I didn't research that very far.

Instead, now I draw the whole event track (and really the whole thing, not
just the visible part), but only once, and cache it in an image.  This
means the initial draw costs more, but after that scrolling is cheap.  The
artifacts are all gone since there is no incremental drawing, and I can draw
the selections afterwards.

This means zooming is now slower, since it has to redraw all the tracks,
even the nonvisible parts.  I can try to get around this by rendering in
chunks, but that's just moving back in the direction of incremental drawing.
Still, maybe it's easier to get right if I do it in larger chunks?  In any
case, I probably scroll more than I zoom, so I'll try it out to see how I
like it.  On the plus side, scrolling is now quick, and I can remove
complicated incremental rendering code.

Theoretically, it should be possible to draw the whole window every time,
in the immediate-mode style.  But fltk is not that kind of library, and
I might even have to re-implement text rendering, which is definitely way
too much work.  GPU acceleration seems necessary in any case.