change Update.CmdUpdate to collect damaged blocks and views
I also changed it from a list to a single monoidal value.  This is more
efficient during diff because it can only diff the the modified blocks
and tracks, and not all of them.  It turns out diffing all of them is
quite slow when there are 3000+ tracks.

This also changes Ui.StateT from lazy to strict StateT, and all of
the Update fields to strict instead of just most of them, if that makes a
difference.