make Internal.add_stack_frame skip Call frames
I also added the 'log-seed' transformer to help with freezing seeds.

Calculating the seed shows up high in the profile, and while the profile
is pretty bogus, it does indicate that the seed gets called a lot.  In
fact, it turns out skipping Calls is more convenient, because otherwise
log-seed wouldn't work, since its very presence would change the seed it's
trying to log.

The annoyance with this is that it breaks all randomness, which means I
have to update all scores, and I really want to stop doing that.