switch seed from CRC32 to just a sum
It turns out CRC32 is kind of slow, probably not the algorithm itself, but
the fact that I have to get it to a pointer and call C, which in practice
seems to slow things down.  Perhaps because the types are small (Doubles
and Chars), and you can't inline across an FFI call.

I looked into Hashable, which was faster, but also explicitly not stable.
So I added Util.Seed, which is like Hashable, but explicitly stable,
because it's just a sum.