repl: clean up some functions, move some things around
Originally I intended to augment the repl with an async notify from the
server, but it just got too complicated and hairy.  After multiple
iterations, none of which could work, I decided that the old mechanism of
tailing seq.log for the quit message is in fact the better way to do it.
These are the refactors left over.