put a marker when an event ends with trailing space
I had this feature long long ago but at some point it disappeared.
Now that I think about it, I may have removed it when I switched to
using the floating text entry, because trailing space is uncommon now.
But you can still get it with method edit, and it's still useful to
tell the difference between "i " and "i" for that reason.