im: change thru protocol from osc to a custom one
This was to fix a problem where it couldn't tell between a single message
with multiple samples (which should overlap), and just separate messages
(which shouldn't).

OSC is inflexible and complicated and wasn't really giving me any benefit.
It does have a "bundle" mechanism that should be able to handle it, but
it's just more complication.

Now I can get rid of both liblo and hosc and various hacks they engender.