solkattu: emphasize on-akshara with bold+red and strokes with just bold
My terminal was set to make bold also red, which meant just red for
emphasized stroke was too close to on-akshara.  So I changed the terminal
so bold is just bold, and recreated the previous look for on-akshara
since I'm used to it, and now I can use uncolored bold for stroke
emphasis.