add Implicit style skeleton
This means the skeleton is determined by the track titles, and is no longer
manually configured.  This in turn means a bunch of stuff that used to
manipulate the skeleton no longer needs to do so.  I haven't deleted that
stuff because I'm not convinced the Implicit skeleton will always be flexible
enough, but hopefully it's good enough for most cases.