nonfloating: replace the floating input window with an overlapping widget
Stuff that used to be done by a FloatingInput containing a WrappedInput is now
done only by WrappedInput.  The reason is that the delicate event dance needed
to get the window to show and disappear at the right time didn't work on X11.
What's more, even if I could get fltk's events to behave sanely (i.e. not emit
multiple unfocus and focus events on a new window), I don't see how it could
ever work with a point-to-focus setup.

That surprised me because I thought I tested this on linux before?  How did I
not notice the inability to enter any text?

This is also really showing the problems in fltk, from things like only having
callbacks, and not allowing any usable argument for them, to more serious
issues like mysterious undocumented and possibly buggy event delivery.  Or
maybe it's just that everything in C++ winds up being this way.

One nice side-effect is that since I'm not making a whole new window anymore,
click to focus and set cursor position simultaneously now works.