integrate: add a separate Score.event_integrated, fix many bugs
Also, I no longer include the set of supported attributes in the CallMap,
though I left the code since I may change my mind.

I also removed `default_scale` which would cause an instrument to set the
scale.  It was confusing and if I want a scale I'll say so!