implement Parse.Ky.merge_instruments
From the comment:

This is an awkward hybrid.  Firstly, the name is inconsistent, internally
they are called "allocation" because "instrument" usually corresponds to
'ScoreT.Instrument', while the allocation is the thing tying instrument to
patch.  But at the UI level I find "allocation" too vague and use
"instrument" for both concepts.

But the main thing is that Allocation corresponds only to a subset of the
actual 'Ui.UiConfig.Allocation'.  The real source of truth is
UiConfig.Allocation.  So the ky equivalent has to merge in any changes that
may have happened to the config (done in Parse.Ky.merge_instruments), which
means automatically updating source, which is fiddly as usual.

A more traditional and simpler way would be to make the ky source be the
source of truth, but that would mean having it serialize all the fields.  I
got pretty far down this route, but it got pretty complicated because
there's a whole new expression language and serialization layer, and I
become unsure if it was really a good idea.

So we do complicated merging for now.