generalize signature parser
This makes it so that the default value doesn't have to be the same type as
the parsed value.  The advantage is that I can coerce many things to
TypedFunction, which is unshowable, but is coerced from normal showable
Vals.

The disadvantages are:

- I lose type safety since there is nothing to enforce that the default
value is in fact convertable to the expected type.  This is because the
default value now goes through a Val and a typecheck, rather than just
directly being the value.  To get it back, I'd need to have promote
Typecheck.from_val to the type level so I could statically assert that
the default is convertable.  Duplicating every Typecheck instance in a
'Coerce a b | a -> b' just to add as a constraint could do it, but of course
it would be nicer to reuse all the typecheck instances.  But, perhaps it's
not such a big deal because a mistyped default will throw an error the first
time it gets tested.

- I have to put type annotations on ambiguous defaults because they are no
longer fixed by the return type.  But since they all go through Val, which
forgets the exact type, if I have a Nothing or [] I don't have to care about
what it's a Maybe or list of, since it will wind coercing to whatever is
needed, so I added Sig.Dummy to signify a don't care type.