remove the separate control signals and pitch signals namespaces
They are now merged into state_environ.  Since I can no longer have
an argument `x=%x` since both `x`s now live in the same namespace,
I also had to convert fully over to Typecheck resolving ControlRefs,
rather than doing it manually.  However, this also means that controls
are finally unified with scalar numbers, in that a scalar will be coerced
into a constant signal, but a signal will also be coerced into a constant
at the time of the call.

This is a major breaking change because the `%x=1` syntax is now gone,
replaced with `x=1`.  I'll have to write a tool to upgrade scores.