remove Score.event_controls and Score.event_pitches
This merges them into event_environ like with derive.  It doesn't
affect the surface syntax, but it should be less error prone because
it's fewer redundant places for data to hide.

There is some awkwardness where VNums that lived in EnvKey are now
just another name for controls, but I expect to unify those up in time.